+++ test/unit/strategy-spec.js
@@ -13,12 +13,14 @@ describe('Strategy', function () {
   });
 
   describe('#buildQuery()', function () {
+    var strategy;
+
     function sharedExamples() {
       context("and given text matches to the function's result", function () {
         it('should return a Query object', function () {
-          var result = this.strategy.buildQuery('@hello');
+          var result = strategy.buildQuery('@hello');
           assert.ok(result instanceof Query);
-          assert.strictEqual(result.strategy, this.strategy);
+          assert.strictEqual(result.strategy, strategy);
           assert.strictEqual(result.term, 'hello');
           assert.ok(Array.isArray(result.match));
         });
@@ -26,31 +28,70 @@ describe('Strategy', function () {
 
       context("and given text does not match to the function's result", function () {
         it('should return null', function () {
-          assert.equal(this.strategy.buildQuery('hello'), null);
+          assert.equal(strategy.buildQuery('hello'), null);
         });
       });
     }
 
     context('when match param is a function' , function () {
       beforeEach(function () {
-        var object = {
-          match: function () { return /(^|\s)@(\w+)$/; },
-        };
-        this.strategy = new Strategy(object);
+        strategy = new Strategy({ match: function () { return /(^|\s)@(\w+)$/; } });
       });
 
-      sharedExamples.call(this);
+      sharedExamples();
     });
 
     context('when match param is a regexp', function () {
       beforeEach(function () {
-        var object = {
-          match: /(^|\s)@(\w+)$/,
-        };
-        this.strategy = new Strategy(object);
+        strategy = new Strategy({ match: /(^|\s)@(\w+)$/ });
       });
 
-      sharedExamples.call(this);
+      sharedExamples();
+    });
+
+    context('when context prop is given', function () {
+      var props;
+
+      beforeEach(function () {
+        props = { match: /(^|\s)@(\w+)$/ };
+        strategy = new Strategy(props);
+      });
+
+      it('should call the function', function () {
+        var spy = props.context = this.sinon.spy();
+        var text = '@hello';
+        strategy.buildQuery(text);
+        assert(spy.calledOnce);
+        assert(spy.calledWith(text));
+      });
+
+      context('and it returns false', function () {
+        beforeEach(function () {
+          props.context = this.sinon.spy(function () {
+            return false;
+          });
+        });
+
+        it('should return null even if the text matches', function () {
+          assert.equal(strategy.buildQuery('@hello'), null);
+        });
+      });
+
+      context('and it returns a string', function () {
+        var returnedString;
+        beforeEach(function () {
+          returnedString = '@world';
+          props.context = this.sinon.spy(function () {
+            return returnedString;
+          });
+        });
+
+        it('should test match with the returned string', function () {
+          var text = '@hello';
+          var result = strategy.buildQuery(text);
+          assert.strictEqual(result.term, 'world');
+        });
+      });
     });
   });